Fix compilation

+++ src/lib/sample.cc
@@ -2,7 +2,7 @@
 
 namespace template_repository {
 
-int64_t add(const int64_t a, const int64_t b) { return a + b; }
-int64_t mul(const int64_t a, const int64_t b) { return a * b; }
+int add(const int a, const int b) { return a + b; }
+int mul(const int a, const int b) { return a * b; }
 
 }  // namespace template_repository
